fix(deviceidentity): #1904 post-merge — device-mTLS material 经 cellmodules helper 收口 (CELLTLS-MATERIAL-FUNNEL-01)
merge origin/develop 引入的 CELLTLS-MATERIAL-FUNNEL-01（#2263）限制 tlsutil.NewServerMTLSConfig
production 调用方为 celltls + adapters/grpc。device-mTLS(renew) listener 是第三个合法 mTLS server
站点（设备客户端证书验证），按 funnel 意图把 mTLS material 构造从 composition root 下沉到
cellmodules helper：

- 新增 cellmodules/deviceidentity.NewDeviceMTLSServerConfig（+ ephemeral cert helper），与 celltls 对称
- cmd/corebundle/deviceidentity_wiring.go 委托该 helper（去掉直接 tlsutil/crypto 调用与 import）
- CELLTLS-MATERIAL-FUNNEL-01 allowlist + godoc + 精度表加 cellmodules/deviceidentity（NewServerMTLSConfig）
- regenerate（与 merge 后 #2429/#2365 等生成器对齐，verify generated 529 files 通过）
